+++ include/Types.h
@@ -5,8 +5,8 @@
 #ifndef DOMINATINGSET_TYPES_H
 #define DOMINATINGSET_TYPES_H
 
-typedef long int v_T;       //type of vertices
-typedef long int e_T;       //type of the label to edge
-typedef long long int w_T;  //weight
+typedef int v_T;       //type of vertices
+typedef int e_T;       //type of the label to edge
+typedef int w_T;       //weight
 
 #endif //DOMINATINGSET_TYPES_H
